legacy-server/requirements.txt
alembic==1.6.5
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
flask-cors==3.0.10
flask-jwt-extended==4.2.3
flask-migrate==3.0.1
flask-sqlalchemy==2.5.1
flask==2.0.1
greenlet==1.1.0
gunicorn==20.1.0
idna==2.10
itsdangerous==2.0.1
jinja2==3.0.1
mako==1.1.4
markupsafe==2.0.1
psycopg2-binary==2.9.1
pyjwt==2.1.0
python-dateutil==2.8.1
python-dotenv==0.18.0
python-editor==1.0.4
python-http-client==3.3.2
redis==3.5.3
requests==2.25.1
sendgrid==6.7.1
six==1.16.0
snowcloud==0.1.2
sqlalchemy==1.4.20
starkbank-ecdsa==1.1.1
urllib3==1.26.6
werkzeug==2.0.1